+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 3.1.0 - TBD
 
 **🌟 New**
-- `Element.prepend` is added.
+- `Element.prepend`, `Element.append` are added.
 
 **🔧 Changes**
 - Bold/Italic texts can not start/end with a space.